fixed some build errors
