+++ app/[lang]/(board)/board/[slug]/page.tsx
@@ -7,9 +7,8 @@ import { Job } from "@/types/job-schema";
 import Image from "next/image";
 import { Input } from "@/components/ui/input";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
-import { Bookmark, MapPin, Share2, X, AlignEndHorizontal, Building2} from "lucide-react"; // Assuming you're using Lucide for icons
+import { Bookmark, MapPin, Share2, X, AlignEndHorizontal} from "lucide-react";
 import { EmbloySpacer } from "@/components/ui/stuff";
-import { FaPhone, FaAt, FaLink, FaFacebook, FaInstagram, FaLinkedin, FaGithub, FaTwitter } from 'react-icons/fa';
 import { cn, replaceNumberWithString } from "@/lib/utils";
 import { Button, buttonVariants } from "@/components/ui/button";
 import Link from "next/link";
@@ -17,11 +16,12 @@ import { toast } from "@/components/ui/use-toast";
 import { siteConfig } from "@/config/site";
 import parse, { domToReact } from 'html-react-parser';
 import { getCurrentUser, User } from "@/lib/api/session";
+import { Company, JobLi, JobParagraph, JobStrong, JobTitle, JobUl, Socials, Stats } from "./utils";
 
 function JobItem({ params, job, user }) {
     const [dict, setDict] = useState<Record<string, any> | null>(null);
     const [shareDropdownOpen, setShareDropdownOpen] = useState(false);
-    const dropdownRef = useRef(null);
+    const dropdownRef = useRef<HTMLDivElement>(null);
 
     useEffect(() => {
         const fetchDictionary = async () => {
@@ -222,8 +222,8 @@ function JobList({ params, jobs, excludeHeader, excludeFooter, user }) {
     };
 
     return (
-        <div className="flex w-full flex-col xl:flex-row items-start justify-between gap-6">
-            <div className="xl:hidden p-4 w-full flex flex-col items-start justify-start gap-6 rounded-lg bg-secondary">
+        <div className="flex w-full flex-col items-start justify-between gap-6 xl:flex-row">
+            <div className="flex w-full flex-col items-start justify-start gap-6 rounded-lg bg-secondary p-4 xl:hidden">
                 <div className="flex flex-col items-start justify-start gap-2">
                     <h1 className="text-left font-heading text-xl">
                         {replaceNumberWithString(dict?.board.list.found, filteredJobs.length.toString())}
@@ -403,285 +403,8 @@ function JobList({ params, jobs, excludeHeader, excludeFooter, user }) {
     );
 }
 
-export function Stats ({ dict, company, className }) {
-    if (company?.company_industry) {
-        return (
-            <>
-            <div className="flex flex-wrap items-start justify-start gap-2">
-                <div className={cn("flex flex-row items-center rounded-full border px-2 dark:border-background dark:text-muted-foreground", className)}>
-                    <h1 className="flex flex-row items-center justify-start gap-1.5 text-xs">
-                        <Building2 className="size-3" />
-                        {company.company_industry}
-                    </h1>
-                </div>
-            </div>
-            <EmbloySpacer className={"h-[4px]"} />
-            </>
-        );} 
-    return null;
-}
 
-export function Socials({ dict, company }) {
-  const facebookRegex = /^(https?:\/\/)?(www\.)?facebook\.com\/[A-Za-z0-9_.-]+\/?$/;
-  const githubRegex = /^(https?:\/\/)?(www\.)?github\.com\/[A-Za-z0-9_.-]+\/?$/;
-  const linkedinRegex = /^(https?:\/\/)?(www\.)?linkedin\.com\/(company|in)\/[A-Za-z0-9_-]+\/?$/;
-  const instagramRegex = /^(https?:\/\/)?(www\.)?instagram\.com\/[A-Za-z0-9_.-]+\/?$/;
-  const twitterRegex = /^(https?:\/\/)?(www\.)?(twitter\.com|x\.com)\/[A-Za-z0-9_.-]+\/?$/;
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const urls = company?.company_urls;
-  const categorized = {
-    facebook: [] as string[],
-    github: [] as string[],
-    linkedin: [] as string[],
-    instagram: [] as string[],
-    twitter: [] as string[],
-    other_urls: [] as string[],
-  };
-
-  urls?.forEach((url) => {
-    if (facebookRegex.test(url)) {
-      categorized.facebook.push(url);
-    } else if (githubRegex.test(url)) {
-      categorized.github.push(url);
-    } else if (linkedinRegex.test(url)) {
-      categorized.linkedin.push(url);
-    } else if (instagramRegex.test(url)) {
-      categorized.instagram.push(url);
-    } else if (twitterRegex.test(url)) {
-      categorized.twitter.push(url);
-    } else {
-      categorized.other_urls.push(url);
-    }
-  });
-
-  return (
-    <>
-      <div className="flex flex-wrap items-start justify-start gap-2 md:flex">
-        {company?.company_phone && (
-          <button
-            onClick={() => {
-              navigator.clipboard.writeText(company.company_phone);
-              return toast({
-                title: replaceNumberWithString(dict?.board.list.copied, "Phone"),
-                variant: "default",
-              });
-            }}
-            className="mx-1 cursor-copy block md:hidden"
-          >
-            <FaPhone className="text-gray-800 dark:text-gray-200" />
-          </button>
-        )}
-        {company?.company_email && (
-          <button
-            onClick={() => {
-              navigator.clipboard.writeText(company.company_email);
-              return toast({
-                title: replaceNumberWithString(dict?.board.list.copied, "Email"),
-                variant: "default",
-              });
-            }}
-            className="mx-1 cursor-copy block md:hidden"
-          >
-            <FaAt className="text-gray-800 dark:text-gray-200" />
-          </button>
-        )}
-        <Button
-            onClick={() => setIsModalOpen(true)}
-            className={cn(buttonVariants({ variant: "transparent", size: "transparent" }), "block md:hidden text-xs underline text-primary-background p-0 h-fit")}
-        >
-            {dict?.board.list.more}
-        </Button>
- 
-
-        <div className="hidden md:flex flex-wrap items-start justify-start gap-2">
-            {company?.company_phone && (
-            <button
-                onClick={() => {
-                navigator.clipboard.writeText(company.company_phone);
-                return toast({
-                    title: replaceNumberWithString(dict?.board.list.copied, "Phone"),
-                    variant: "default",
-                });
-                }}
-                className="mx-1 flex flex-row items-center gap-0.5 break-all cursor-copy"
-            >
-                <FaPhone className="text-gray-800 dark:text-gray-200 text-xs" />
-                <p className="text-xs">{company.company_phone}</p>
-            </button>
-            )}
-            {company?.company_email && (
-            <button
-                onClick={() => {
-                navigator.clipboard.writeText(company.company_email);
-                return toast({
-                    title: replaceNumberWithString(dict?.board.list.copied, "Email"),
-                    variant: "default",
-                });
-                }}
-                className="mx-1 flex flex-row gap-0.5 items-center break-all cursor-copy"
-            >
-                <FaAt className="text-gray-800 dark:text-gray-200 text-xs" />
-                <p className="text-xs">{company.company_email}</p>
-            </button>
-            )}
-          {categorized.facebook.length > 0 &&
-            categorized.facebook.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaFacebook className="text-blue-600 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-          {categorized.instagram.length > 0 &&
-            categorized.instagram.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaInstagram className="text-pink-500 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-          {categorized.linkedin.length > 0 &&
-            categorized.linkedin.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaLinkedin className="text-blue-700 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-          {categorized.github.length > 0 &&
-            categorized.github.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaGithub className="text-gray-800 dark:text-gray-200 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-          {categorized.twitter.length > 0 &&
-            categorized.twitter.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaTwitter className="text-blue-400 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-          {categorized.other_urls.length > 0 &&
-            categorized.other_urls.map((url) => (
-              <a
-                href={url}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="mx-1 flex flex-row gap-0.5 break-all items-center"
-              >
-                <FaLink className="text-gray-800 dark:text-gray-200 text-xs" />
-                <p className="text-xs">{url}</p>
-              </a>
-            ))}
-        </div>
-      </div>
-
-      {isModalOpen && (
-        <div className="fixed inset-0 flex items-center justify-center bg-gray-800 bg-opacity-75 z-50">
-          <div className="bg-white dark:bg-gray-900 p-6 rounded-md shadow-md w-11/12 md:w-1/2">
-            <div className="flex flex-row justify-between items-start">
-                <h2 className="text-xl font-semibold mb-4">Links</h2>
-                <button
-                    onClick={() => setIsModalOpen(false)}
-                    className=" text-gray-800 dark:text-gray-200 text-xl font-heading"
-                >
-                    &times;
-                </button>
-            </div>
-            
-            <div className="flex flex-col gap-2">
-              {Object.keys(categorized).map((key) => {
-                if (categorized[key].length > 0) {
-                  return (
-                    <div key={key} className="flex flex-col gap-2">
-                        {categorized[key].map((url) => (
-                          <a
-                            key={url}
-                            href={url}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            className="flex items-center w-fit flex flex-row gap-1.5 items-center"
-                          >
-                            <FaLink className="text-gray-800 dark:text-gray-200 text-xs" />
-                            <p className="text-xs">{url}</p>
-                          </a>
-                        ))}
-                    </div>
-                  );
-                }
-              })}
-            </div>
-          </div>
-        </div>
-      )}
-    </>
-  );
-}
-
-export const JobTitle = ({ children }) => (
-    <h1 className="font-heading text-xl ">{children}</h1>
-);
-export const JobParagraph = ({ children }) => (
-    <p className="text-base">{children}</p>
-);
-export const JobUl = ({ children }) => (
-    <ul className="ml-8 list-disc text-base">{children}</ul>
-);
-  
-export const JobLi = ({ children }) => (
-    <li className="text-base">{children}</li>
-);
-  
-export const JobStrong = ({ children }) => (
-    <strong className="font-heading text-base">{children}</strong>
-);
-export interface CompanyDescription {
-    id: number;
-    name: string;
-    body: string;
-    record_type: string;
-    record_id: number;
-    created_at: string;
-    updated_at: string;
-}
-  
-export interface Company {
-    id: number;
-    company_name: string;
-    company_phone: string;
-    company_email: string;
-    company_urls: string[];
-    company_industry: string;
-    company_description: CompanyDescription;
-    company_logo: string;
-    company_slug: string;
-}
-export default function Page({ params, excludeHeader, excludeFooter }) {
+export default function Page({ params}) {
     const [user, setUser] = useState<User | null>(null)
     const [shareDropdownOpen, setShareDropdownOpen] = useState(false);
     const dropdownRef = useRef(null);
